fix ep nb in title + regex
The folder name is used as serie name and removed from the filename to reduce complexity. It use to skip non regex logic and relying on the regex when the episode number is contained in the title (scrubbing years beforehand) and i broke the regex. i changed the logic so it uses the non regex engine for that and also check early for "serie - xxx - title.ext" naming convention so that should fix stuff. also reged used to trigger on fifles containing op or ed in titles, not good...