Update Absolute Series Scanner.py
https://forums.plex.tv/discussion/comment/1309353/#Comment_1309353
Running into a new anidb2 mapping oddity. This time it is Strike Witches, specifically the movie. In ScudLee's file it is correctly set to map the movie "episode 01" to s0e02, but ASS keeps putting it to s0e01 and I'm not sure why. I wanted to run it by you though in case there is a bug.