YouTube API key loading from agent settings to avoid "Daily Limit Exceeded"
https://github.com/ZeroQI/Absolute-Series-Scanner/issues/166
On line 620 of Absolute Series Scanner.py, the API key for the youtube client is hardcoded, even though it's possible to set it through the UI settings within plex. This has caused the YouTube scanner to stop working properly as the API is currently returning "Daily Limit Exceeded", causing the scanner to crash when trying to run line 631 Log.info('---- count: {}'.format(len(json_full['items'])))

If this could be updated to use the API key entered in the plex settings, that would be much better as it would allow us to avoid these issues in future.