Fix filename cleaning of folder name
If the folder name was short enough it would incorrectly remove all instances of that string from the filename. It was also not checking that the name was also found as an actual prefix at the start of the string.
EX: ".../Angel/Angel - S00E502 - Introducing Angel" turned into an episode name of "Introducing"