Fixes a bug in specials ordering for forced 'anidb'
If specials contained an alphabetical character (i.e. 2a / 2b), the script would perform a string / integer comparison which could throw off the `cumulative_offset` (for example, `2 < '1' == true`).

This fixes this so that the variable `ep` is always an integer value when being compared in the initialisation of `cumulative_offset`.