Youtube2 functionality, forcing MMDDhhmm Episode #
Reintroduce the Youtube2 switch/source that will fabricate an Episode Number for the channels which uses the switch ( [youtube2-UC...] ) in the format of MMDDhhmm (Month, Day, Hour, Minute). This will provide a consistent and unique episode numbering system which will allow the Youtube Meta Agent to correctly pull and apply metadata for episodes released on the same day.

It will pull the Month and Day from the filename if present, and then use the file's Modified time for the Hour and Minute to differentiate from other episodes from the same day. If the date is not present in the filename then it will pull all four from the file's Modified time.