Remove "DD2.0" to prevent Ex0 being handled as E0x
When E01 and E10 were present (and E02 and E20, etc.) along with DD2.0
and 1080p the episode number parser would report E10 (and E20, etc.) as
E01 (and E02, etc.).  This caused "Episode 1" to have two versions that
were actually E01 and E10, and the same for "Episode 2" and so on.

This change just follows the pattern of cleaning up predictable parts of
file names that cause problems when extracting the episode number.

Closes #428